[tagger/mock] Refactor to use fake impl directly and avoid unwanted dependencies (#31297)
