+++ comp/process/agent/agentimpl/agent_linux_test.go
@@ -17,6 +17,9 @@ import (
 	"go.uber.org/fx"
 
 	configComp "github.com/DataDog/datadog-agent/comp/core/config"
+	log "github.com/DataDog/datadog-agent/comp/core/log/def"
+	logmock "github.com/DataDog/datadog-agent/comp/core/log/mock"
+	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig/sysprobeconfigimpl"
 	taggerMock "github.com/DataDog/datadog-agent/comp/core/tagger/mock"
 	"github.com/DataDog/datadog-agent/comp/core/telemetry"
 	"github.com/DataDog/datadog-agent/comp/core/telemetry/telemetryimpl"
@@ -129,6 +132,9 @@ func TestProcessAgentComponentOnLinux(t *testing.T) {
 				submitterimpl.MockModule(),
 				taggerMock.Module(),
 				statsd.MockModule(),
+				fx.Provide(func(t testing.TB) log.Component { return logmock.New(t) }),
+				configComp.MockModule(),
+				sysprobeconfigimpl.MockModule(),
 				Module(),
 
 				fx.Replace(configComp.MockParams{Overrides: map[string]interface{}{
@@ -197,6 +203,9 @@ func TestStatusProvider(t *testing.T) {
 					"process_config.run_in_core_agent.enabled": true,
 				}}),
 				processcheckimpl.MockModule(),
+				fx.Provide(func(t testing.TB) log.Component { return logmock.New(t) }),
+				configComp.MockModule(),
+				sysprobeconfigimpl.MockModule(),
 				fx.Provide(func() func(c *checkMocks.Check) {
 					return func(c *checkMocks.Check) {
 						c.On("Init", mock.Anything, mock.Anything, mock.AnythingOfType("bool")).Return(nil).Maybe()
@@ -242,6 +251,9 @@ func TestTelemetryCoreAgent(t *testing.T) {
 			"telemetry.enabled":                        true,
 		}}),
 		processcheckimpl.MockModule(),
+		fx.Provide(func(t testing.TB) log.Component { return logmock.New(t) }),
+		configComp.MockModule(),
+		sysprobeconfigimpl.MockModule(),
 		fx.Provide(func() func(c *checkMocks.Check) {
 			return func(c *checkMocks.Check) {
 				c.On("Init", mock.Anything, mock.Anything, mock.AnythingOfType("bool")).Return(nil).Maybe()
